# HW10 Program optimization