--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -1,6 +1,6 @@
 module github.com/derticom/home_work/hw10_program_optimization
 
-go 1.23
+go 1.22
 
 require github.com/stretchr/testify v1.10.0