--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -2,65 +2,36 @@ package hw10programoptimization
 
 import (
 	"encoding/json"
-	"fmt"
 	"io"
-	"regexp"
 	"strings"
 )
 
 type User struct {
-	ID       int
-	Name     string
-	Username string
-	Email    string
-	Phone    string
-	Password string
-	Address  string
+	Email string `json:"Email"` //nolint:tagliatelle // Tag name defined in JSON.
 }
 
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
+	domainStat := DomainStat{}
+	decoder := json.NewDecoder(r)
 
-type users [100_000]User
+	for {
+		user := User{}
 
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := io.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
-		}
-		result[i] = user
-	}
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
-	result := make(DomainStat)
-
-	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
+		if err := decoder.Decode(&user); err != nil {
+			if err == io.EOF {
+				return domainStat, nil
+			}
 			return nil, err
 		}
 
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+		if !strings.HasSuffix(user.Email, "."+domain) {
+			continue
 		}
+
+		num := domainStat[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
+		num++
+		domainStat[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
 	}
-	return result, nil
 }